feat(auth): unified sign-in page for students and landlords
Single /signin page replaces the two role-specific login screens.
After Supabase auth succeeds the page calls /api/auth/me to detect
role and redirects: students → /student/account (or ?next= param),
landlords → /property/thessaloniki/landlord/dashboard, orphaned
accounts → friendly error with signup links.

Old URLs kept working via server-side redirect stubs:
  /student/login → /signin (preserves ?next=, ?email=)
  /property/<city>/landlord/login → /signin (preserves ?email=)
  /property/<city>/landlord/login (non-locale) → /signin directly

Navbar now shows a single "Sign in" link instead of the previous
student + landlord pair.

Signup role-conflict CTAs on both signup pages now deep-link to
/signin?email=... instead of the old role-specific login paths.
